refactor: se mueve logica servicio, se configura el proyecto para que al momento del deploy se utilizen variables de entorno para configuraciones de aplicacion
