--- AWSServerless/Options.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AWSServerless
-{
-    public class Options
-    {
-        public string AWS_SPEECH_REGION { get; set; }
-        public string AWS_SPEECH_ACCESS_KEY { get; set; }
-        public string AWS_SPEECH_SECRET_KEY { get; set; }
-    }
-}
-
