+++ AWSServerless/Properties/launchSettings.json
@@ -11,15 +11,24 @@
     "IIS Express": {
       "commandName": "IISExpress",
       "launchBrowser": true,
+      "launchUrl": "swagger",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        "speechRegion": "Valor",
+        "swaggerActive": "true",
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "speechAcessKey": "Valor",
+        "speechSecretKey": "Valor"
       }
     },
     "AWSServerless": {
       "commandName": "Project",
       "launchBrowser": true,
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        "speechRegion": "Valor",
+        "swaggerActive": "true",
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "speechAcessKey": "Valor",
+        "speechSecretKey": "Valor"
       },
       "applicationUrl": "https://localhost:5001;http://localhost:5000"
     },
@@ -34,8 +43,12 @@
       "launchBrowser": true,
       "launchUrl": "https://localhost:5001",
       "environmentVariables": {
+        "speechRegion": "Valor",
+        "swaggerActive": "true",
         "ASPNETCORE_URLS": "https://localhost:5001;http://localhost:5000",
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "speechAcessKey": "Valor",
+        "speechSecretKey": "Valor"
       },
       "distributionName": ""
     }
