+++ AWSServerless/Startup.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Application.IServices;
+using Application.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -25,11 +27,9 @@ namespace AWSServerless
             .SetBasePath(env.ContentRootPath)
             .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
             .AddConfiguration(configuration)
-            .AddUserSecrets<Options>()
             .AddEnvironmentVariables();
 
             Configuration = builder.Build();
-
         }
 
         public static IConfiguration Configuration { get; private set; }
@@ -37,12 +37,7 @@ namespace AWSServerless
         // This method gets called by the runtime. Use this method to add services to the container
         public void ConfigureServices(IServiceCollection services)
         {
-#pragma warning disable ASP0000 // Do not call 'IServiceCollection.BuildServiceProvider' in 'ConfigureServices'
-            services.Configure<Options>(Configuration.GetSection(nameof(Options)))
-                .AddOptions()
-                .BuildServiceProvider();
-#pragma warning restore ASP0000 // Do not call 'IServiceCollection.BuildServiceProvider' in 'ConfigureServices'
-
+            services.AddScoped<IPresignedUrlService, PresignedUrlService>();
             services.AddControllers();
             services.AddSwaggerGen(c =>
             {
@@ -72,8 +67,6 @@ namespace AWSServerless
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-                app.UseSwagger();
-                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "SignedURL v1"));
             }
             else
             {
@@ -82,6 +75,20 @@ namespace AWSServerless
                 app.UseHsts();
             }
 
+            string swaggerActive = Environment.GetEnvironmentVariable("swaggerActive");
+            bool flag = bool.Parse(swaggerActive);
+
+            string auxUri = "";
+
+            if (flag.Equals(true))
+            {
+                app.UseSwagger();
+                if (env.IsProduction())
+                {
+                    auxUri = "Prod";
+                }
+                app.UseSwaggerUI(c => c.SwaggerEndpoint(@$"{auxUri}/swagger/v1/swagger.json", "AWSServerless v1"));
+            }
 
             app.UseHttpsRedirection();
             app.UseStaticFiles();
@@ -99,7 +106,147 @@ namespace AWSServerless
                         pattern: "{controller}/{action=Index}/{id?}");
                 endpoints.MapGet("/", async context =>
                 {
-                    await context.Response.WriteAsync("Welcome to running ASP.NET Core on AWS Lambda");
+                    string html = $@"<!doctype html>
+<html lang='en'>
+
+<head>
+    <!-- Required meta tags -->
+    <meta charset='utf-8'>
+    <meta name='viewport' content='width=device-width, initial-scale=1'>
+
+    <!-- Bootstrap CSS -->
+    <link href='https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css' rel='stylesheet'
+        integrity='sha384-F3w7mX95PdgyTmZZMECAngseQB83DfGTowi0iMjiWaeVhAn4FJkqJByhZMI3AhiU' crossorigin='anonymous'>
+    <link href='https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css' rel='stylesheet'>
+
+    <link rel='stylesheet' href='https://unpkg.com/@coreui/icons/css/all.min.css'>
+
+
+    <title>Hello, world!</title>
+</head>
+
+<body>
+
+    <div class='px-4 py-5 my-5 text-center'>
+        <img class='d-block mx-auto mb-4' src='	https://getbootstrap.com/docs/5.1/assets/brand/bootstrap-logo.svg' alt='' width='72' height='57'>
+        <h1 class='display-5 fw-bold'>Welcome to running ASP.NET Core on AWS Lambda</h1>
+        <div class='col-lg-6 mx-auto'>
+            <p class='lead mb-4'>Quickly design and customize responsive mobile-first sites with Bootstrap, the world’s
+                most popular front-end open source toolkit, featuring Sass variables and mixins, responsive grid system,
+                extensive prebuilt components, and powerful JavaScript plugins.</p>
+        </div>
+    </div>
+
+
+
+    <div class=''>
+        <div class='container'>
+            <div class='row'>
+
+                <div class='col-lg-4 col-xs-12 text-center mb-3'>
+                    <div class=''>
+                        <i class='fa fa-github fa-3x' aria-hidden='true'></i>
+                        <div class=''>
+                            <h3>Github</h3>
+                        </div>
+                        <div class=''>
+                            <span>Lorem ipsum dolor sit amet, id quo eruditi eloquentiam. Assum decore te sed. Elitr
+                                scripta ocurreret qui ad.</span>
+                        </div>
+                        <div class=''>
+                            <a href='https://github.com/patricioarena/ProgramacionEnTiempoReal'>Learn More</a>
+                        </div>
+                    </div>
+                </div>
+
+                <div class='col-lg-4 col-xs-12 text-center mb-3'>
+                    <div class=''>
+                        <img src='https://petstore.swagger.io/favicon-32x32.png' width='50' height='50'>
+                        <div class=''>
+                            <h3>Swagger</h3>
+                        </div>
+                        <div class=''>
+                            <span>Lorem ipsum dolor sit amet, id quo eruditi eloquentiam. Assum decore te sed. Elitr
+                                scripta ocurreret qui ad.</span>
+                        </div>
+                        <div class=''>
+                            <a href='{auxUri}/swagger/index.html'>Learn More</a>
+                        </div>
+                    </div>
+                </div>
+
+                <div class='col-lg-4 col-xs-12  text-center mb-3'>
+                    <div class=''>
+                        <i class='fa fa-twitter fa-3x' aria-hidden='true'></i>
+                        <div class=''>
+                            <h3>Twitter</h3>
+                        </div>
+                        <div class=''>
+                            <span>Lorem ipsum dolor sit amet, id quo eruditi eloquentiam. Assum decore te sed. Elitr
+                                scripta ocurreret qui ad.</span>
+                        </div>
+                        <div class=''>
+                            <a href='#'>Learn More</a>
+                        </div>
+                    </div>
+                </div>
+
+                <div class='col-lg-4 col-xs-12 text-center mb-3'>
+                    <div class=''>
+                        <i class='fa fa-facebook fa-3x' aria-hidden='true'></i>
+                        <div class=''>
+                            <h3>Facebook</h3>
+                        </div>
+                        <div class=''>
+                            <span>Lorem ipsum dolor sit amet, id quo eruditi eloquentiam. Assum decore te sed. Elitr
+                                scripta ocurreret qui ad.</span>
+                        </div>
+                        <div class=''>
+                            <a href='#'>Learn More</a>
+                        </div>
+                    </div>
+                </div>
+
+                <div class='col-lg-4 col-xs-12 text-center mb-3'>
+                    <div class=''>
+                        <i class='fa fa-pinterest-p fa-3x' aria-hidden='true'></i>
+                        <div class=''>
+                            <h3>Pinterest</h3>
+                        </div>
+                        <div class=''>
+                            <span>Lorem ipsum dolor sit amet, id quo eruditi eloquentiam. Assum decore te sed. Elitr
+                                scripta ocurreret qui ad.</span>
+                        </div>
+                        <div class=''>
+                            <a href='#'>Learn More</a>
+                        </div>
+                    </div>
+                </div>
+
+                <div class='col-lg-4 col-xs-12 text-center mb-3'>
+                    <div class=''>
+                        <i class='fa fa-google-plus fa-3x' aria-hidden='true'></i>
+                        <div class=''>
+                            <h3>Google</h3>
+                        </div>
+                        <div class=''>
+                            <span>Lorem ipsum dolor sit amet, id quo eruditi eloquentiam. Assum decore te sed. Elitr
+                                scripta ocurreret qui ad.</span>
+                        </div>
+                        <div class=''>
+                            <a href='#'>Learn More</a>
+                        </div>
+                    </div>
+                </div>
+
+            </div>
+        </div>
+    </div>
+</body>
+
+</html>";
+
+                    await context.Response.WriteAsync(html);
                 });
             });
         }
