+++ AWSServerless/appsettings.json
@@ -5,15 +5,18 @@
     }
   },
   "Lambda.Logging": {
-    "IncludeCategory": false,
-    "IncludeLogLevel": false,
+    "IncludeCategory": true,
+    "IncludeLogLevel": true,
     "IncludeNewline": true,
     "IncludeException": true,
     "IncludeEventId": true,
     "IncludeScopes": true,
     "LogLevel": {
-      "Default": "Debug",
-      "Microsoft": "Information"
+      "Default": "Information",
+      "System": "Warning",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Warning"
     }
-  }
+  },
+  "AllowedHosts": "*"
 }
