+++ AWSServerless/aws-lambda-tools-defaults.json
@@ -1,15 +1,18 @@
+
 {
-  "Information": [
-    "This file provides default values for the deployment wizard inside Visual Studio and the AWS Lambda commands added to the .NET Core CLI.",
-    "To learn more about the Lambda commands with the .NET Core CLI execute the following command at the command line in the project root directory.",
-    "dotnet lambda help",
-    "All the command line options for the Lambda command can be specified in this file."
-  ],
-  "profile": "",
-  "region": "",
-  "configuration": "Release",
-  "framework": "netcoreapp3.1",
-  "s3-prefix": "AWSServerless/",
-  "template": "serverless.template",
-  "template-parameters": ""
+    "Information" : [
+        "This file provides default values for the deployment wizard inside Visual Studio and the AWS Lambda commands added to the .NET Core CLI.",
+        "To learn more about the Lambda commands with the .NET Core CLI execute the following command at the command line in the project root directory.",
+        "dotnet lambda help",
+        "All the command line options for the Lambda command can be specified in this file."
+    ],
+    "profile"     : "default",
+    "region"      : "us-east-1",
+    "configuration" : "Release",
+    "framework"     : "netcoreapp3.1",
+    "s3-prefix"     : "AWSServerless/",
+    "template"      : "serverless.template",
+    "template-parameters" : "",
+    "s3-bucket"           : "mylambdatestfuntion",
+    "stack-name"          : "myLambdaTestFuntion"
 }
